:bug: fix circular import error
