+++ fastapi_async_langchain/websockets/__init__.py
@@ -1,9 +1,8 @@
-from .base import BaseLangchainWebsocketConnection, Response
+from .base import BaseLangchainWebsocketConnection
 from .llm import LLMChainWebsocketConnection
 from .retrieval_qa import RetrievalQAWebsocketConnection
 
 __all__ = [
-    "Response",
     "BaseLangchainWebsocketConnection",
     "LLMChainWebsocketConnection",
     "RetrievalQAWebsocketConnection",
